Ghostscript has removed pswrite device.
Because some versions of the ps2write device were buggy we now need to detect whether or not the pswrite device is available and act accordingly.

If at some point we learn which versions of Ghostscript have the buggy ps2write device then we could default to ps2write except for those versions.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2335 5505949e-d877-4686-9e03-c53b7a51b376