This adds the JobHostID feature.
This merely prepends job id numbers with any configured JobHostID in all communications with the client.  It does not affect job ids internally, and it does not add them to notification e-mails, etc.  This only deals with the client-server communication.

This is done for a client to be able to distinguish between jobs submitted on one host from jobs submitted on another host by way of the client-viewed job id number.

This may seem a little silly at this point, but the intention is to subsequently develop the ability for hfaxd to act as a proxy-client to other HylaFAX servers when a client attempts to operate on a job which actually exists on a different server.

Consider a "cluster" of HylaFAX servers that all share a single DNS hostname, "fax.mydomain.org".  If a client submits a job to the "clustered" name, how will it know which of the servers in the cluster actually took the job?  Furthermore, if it needs to subsequently delete or modify the job, how will it know which host to contact?  That is the purpose of JobHostID.

Upcoming development in turning hfaxd into a proxy-client for other hosts in the cluster is to permit the client to perform job operations subsequent to the submission on the "cluster" hostname rather than needing to understand the codification of the individual servers in the job id.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2341 5505949e-d877-4686-9e03-c53b7a51b376