The "npages" value carries different meanings internally, depending on where it's used.
When "npages" is used in faxsend it refers to the number of pages sent in that session.

When "npages" is used in the sendq file it refers to the total number of pages of the job sent so far.

So, when dealing with a proxy send situation we need to do both: record the number of pages sent in that session AND add that number to the running total.

This patch also fixes the overwriting of "totpages" by the proxy operation which bug stemmed from the same misunderstanding as "npages".