This is a first-stab at trying to support >256 pages in Class 1 ECM.
T.30A provides for an 8-bit page-count field.  So that maxes-out at 256 pages.

These changes cause the send-side to "roll-over" the count so that the 257th page starts the field over again as if it were the 1st page.  These changes also allow the receive-side to interpret this and count the pages correctly.

This has not been tested against other fax softwares or fax machines.  I am not presently aware of how those deal with >256 pages or cope with receiving >256 pages.  The only experience I have indicates that at least one fax receiver out there does not care about the PPS page-count field at all, and will receive pages regardless of the indication.  Obviously we're not that lax as a receiver.

If we find that other senders behave differently when going >256 pages we may need to change things.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2362 5505949e-d877-4686-9e03-c53b7a51b376