The setDataTimeout() function has had a long-time (since at least v3.0pl0) bug.
The bug completely ignored 7200 bps, and so data transmission timeouts would happen on large transmissions at 7200 bps where the modem buffer allowed for a huge "slurping" of all the data.

Furthermore, the code commenting seemed to be contrary to the code operation as it claimed that the nominal times were for 9600 bps, but then adjusted things from there.

Plus, at 9600 bps the 60-second data timeout was really too slow, anyway, as it was subsequently used for ECM transmissions.

So this fixes that oversight... AND the dataTimeout value is now set dynamically for the amount of page data being transmitted (at least on the first attempt).