DynamicConfig assignment of Class 1 modem data format support has never worked properly.
The pokeConfig() function was properly *adding* support for the various formats, but it was not properly *removing* support for them... which ultimately made them not take effect until the subsequent call handled by the modem.

These changes explicity add *and* remove support for the various data formats (MMR, JBIG, JPEG) as instructed by DynamicConfig.

[This commit was originally made to the Sourceforge SVN repository on Thu Jul 16 2015 05:06:49 GMT+0000 (UTC), however, it was lost by the storage fault at Sourceforge later that day and was not present in the back-up.]



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2375 5505949e-d877-4686-9e03-c53b7a51b376