This makes faxsetup fix blind references in Fontmap.HylaFAX.
Apparently Ghostscript is okay with Fontmap.GS when it references named fonts whose files are not explicitly given elsewhere.  However, this does not work for us.

So, this fix follows an approach similar to what Chris Severance did in "fmfix.pl" file in his archlinux aur "hylafaxplus" repository... but we don't (and can't) use perl.