On second thought, let's actually fix the bug here instead of perpetuate it.
The unused code in the else clause is not unimportant.  ntries refers to pages, and if the tif has more than one page then npages will need to be incremented in sendPhaseB().  However, it's not.

This reverts the last change and fixes the bug by teaching sendPhaseB() to increment npages.