Job priority "bounding" interferes with jobcontrol action.
If a client submits a job as "high" priority but jobcontrol moves it to a "low" priority setting, then this bounding logic stops normal priority escalation on retries.  So, with this "bounding" jobcontrol is limited to moving the job priority to within the undocumented priority bounds, to being content with the priority staying stuck, to managing the priority on every instance, or to moving the priority to "bulk" where bounding is not a factor.

I feel that the change years-ago to reduce the priority of "bulk" jobs on every attempt made job priority bounding irrelevant and near-useless.

And, since it is troublesome for jobcontrol and undocumented outside of the code, itself, I'm removing it.