On very busy systems we can see regular "FIFO Read full" messages.
This change increases the read size to try to reduce the likelihood of such instances, but this change also adds some diagnostic information in the "Read full" instances.

Read full conditions are not problematic as long as the process loops fast-enough to deplete the FIFO buffer before it fills.  Problems arise when the buffer fills.

The larger read size should help some.  However, it may be necessary to increase the FIFO buffer size on extremely busy systems.  The added diagnostic information will help determine when such cases arise.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2424 5505949e-d877-4686-9e03-c53b7a51b376