This reduces the number of JWAIT commands issued (to one).
The only purpose the repeated JWAIT served was to "ping" the server and ensure that it was still responsive.  This behavior came with the added consequence of initiating a trigger for each JWAIT issued, and that led to a lot of trigger consumption and stress on the FIFOs.

The "ping" benefits can be derived equally from repeated ABOR commands, without the consumption of triggers and FIFOs.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2426 5505949e-d877-4686-9e03-c53b7a51b376