This adds CLEANFAXDATA and BADFAXLINES TIFF tags onto ECM-received faxes.
This is useful for diagnostic purposes.  In doing this we equate ECM frames with "lines".

This patch also fixes a bug for non-ECM faxes where they were always being tagged with CLEANFAXDATA_REGENERATED when that was only correct when copy quality checking (and correction) was occurring.  This fixes that and tags CLEANFAXDATA_UNCLEAN when copy quality checking is not being performed.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2428 5505949e-d877-4686-9e03-c53b7a51b376