This adds CLEANFAXDATA and BADFAXLINES TIFF tags onto ECM-received faxes.
This is useful for diagnostic purposes.  In doing this we equate ECM frames with "lines".

This patch also fixes a bug for non-ECM faxes where they were always being tagged with CLEANFAXDATA_REGENERATED when that was only correct when copy quality checking (and correction) was occurring.  This fixes that and tags CLEANFAXDATA_UNCLEAN when copy quality checking is not being performed.