Escaping of double-quotes was not working in particular when the double-quote was the first or last character in the string.
One mistake was that fxStr.next() was being used looing for a zero-value for a true/false indication.  That has to be done with fxStr.nextR().  This mistake completely defeated its intent and caused all strings to be evaluated except those that started with a double-quote.

The second mistake was that starting the existing nextR() at r-1 caused it to ignore the last character in the string.

This change fixes both errors.