A bugfix and enhancement for faxmail.
There was a bug in faxmail's transition between a "message" part and an external part which left the postscript output unprepared for the output from the external handler.  This is fixed now.

There was also possibly a bug in faxmail's arguments to the MIMEHandlers for the "id" and "disposition" arguments.  If so, then this was also fixed in the process of changing the arguments to be of the "-id <info>" format rather than "id:<info>".  Furthermore, the Content-Type name and Content-Disposition filename parameters are also now passed to the MIMEHandlers.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2442 5505949e-d877-4686-9e03-c53b7a51b376