Fix an incongruency in parsing config files.
On one hand we're limited to reading 1024 bytes, but then we open it up later to 2048.  Let's be consistent.

The smaller limit ends up having an effect on very long configuration item lines such as with "FontPath".  If that configuration line exceeds 1024 bytes then whatever comes after 1024 bytes is treated as a new line.  So, in a FontPath listing the path before the truncation will be corrupted, and the remainder of that path, parsed as a new line, will be ignored... AND if a colon (:) is not found later then an error will be displayed.

Ultimately, maybe, we'll need to teach the parser how to not truncate and corrupt these kinds of long configuration items.  For now, though, it works just to increase the buffer size.



git-svn-id: https://svn.code.sf.net/p/hylafax/HylaFAX+@2443 5505949e-d877-4686-9e03-c53b7a51b376